requirements.txt
pillow~=10.3.0
pip~=24.0
wheel~=0.43.0
Django~=5.0.8
sqlparse~=0.4.4
asgiref~=3.8.1
typing_extensions~=4.10.0
certifi~=2024.2.2
telebot~=0.0.5
requests~=2.31.0
idna~=3.7
urllib3~=2.2.1
django-ckeditor~=6.7.1
django-modeltranslation~=0.19.9
python-decouple~=3.8
whitenoise~=6.7.0